Refactor trace store chunk access helpers
